smpp_core/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmpplib==0.6.4
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
constantly==23.10.4
cryptography==42.0.5
executing==2.0.1
Flask==3.0.2
frozenlist==1.4.1
hyperlink==21.0.0
icecream==2.1.3
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
naz==0.8.1
proxy.py==2.4.3
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
smpplib==2.2.3
smpppdu==0.1.2
smppy==0.3.2
twilio==9.0.2
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zope.interface==6.2
